APIs/googleapis.com/displayvideo/v3/python-aiohttp-client/tests/test_targeting_types_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.list_targeting_options_response import ListTargetingOptionsResponse
from openapi_server.models.search_targeting_options_request import SearchTargetingOptionsRequest
from openapi_server.models.search_targeting_options_response import SearchTargetingOptionsResponse
from openapi_server.models.targeting_option import TargetingOption


pytestmark = pytest.mark.asyncio

async def test_displayvideo_targeting_types_targeting_options_get(client):
    """Test case for displayvideo_targeting_types_targeting_options_get

    
    """
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('advertiserId', 'advertiser_id_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/v3/targetingTypes/{targeting_type}/targetingOptions/{targeting_option_id}'.format(targeting_type='targeting_type_example', targeting_option_id='targeting_option_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_displayvideo_targeting_types_targeting_options_list(client):
    """Test case for displayvideo_targeting_types_targeting_options_list

    
    """
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('advertiserId', 'advertiser_id_example'),
                    ('filter', 'filter_example'),
                    ('orderBy', 'order_by_example'),
                    ('pageSize', 56),
                    ('pageToken', 'page_token_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/v3/targetingTypes/{targeting_type}/targetingOptions'.format(targeting_type='targeting_type_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_displayvideo_targeting_types_targeting_options_search(client):
    """Test case for displayvideo_targeting_types_targeting_options_search

    
    """
    body = {"businessChainSearchTerms":{"businessChainQuery":"businessChainQuery","regionQuery":"regionQuery"},"pageSize":0,"pageToken":"pageToken","poiSearchTerms":{"poiQuery":"poiQuery"},"geoRegionSearchTerms":{"geoRegionQuery":"geoRegionQuery"},"advertiserId":"advertiserId"}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/v3/targetingTypes/{targeting_type}/targetingOptions:search'.format(targeting_type='targeting_type_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

